refactor to separate registry root
