+++ src/index.ts
@@ -1,14 +1,14 @@
-import {
-  Web3,
-  HttpClient,
-  ProofPoint,
-  ProofPointIssueResult,
-  ProofPointRegistry,
-  ProofPointValidateResult,
-  ProofPointStatus,
-  ProofPointEventType,
-  ProofPointEvent,
-} from "./proofPoints";
+import Web3 from "web3";
+import { HttpClient } from "./httpClient";
+import { ProofPoint } from "./proofPoint";
+import { ProofPointIssueResult } from "./proofPointIssueResult";
+import { ProofPointRegistry } from "./proofPointRegistry";
+import { ProofPointValidateResult } from "./proofPointValidateResult";
+import { ProofPointStatus } from "./proofPointStatus";
+import { ProofPointEventType } from "./proofPointEventType";
+import { ProofPointEvent } from "./proofPointEvent";
+import { ProofPointRegistryRoot } from "./proofPointRegistryRoot";
+
 import {
   StorageProvider,
   IpfsStorageProvider,
@@ -28,4 +28,5 @@ export {
   ProofPointStatus,
   ProofPointEventType,
   ProofPointEvent,
+  ProofPointRegistryRoot,
 };
