+++ src/proofPointEventType.ts
@@ -1,0 +1,10 @@
+/**
+ * Proof Point event type, the type of an {@link ProofPointEvent}.
+ */
+enum ProofPointEventType {
+  Issued,
+  Committed,
+  Revoked,
+}
+
+export { ProofPointEventType };
