+++ src/proofPointIssueResult.ts
@@ -1,0 +1,9 @@
+import { ProofPoint } from "./proofPoint";
+
+interface ProofPointIssueResult {
+  proofPointId: string;
+  transactionHash: string;
+  proofPointObject: ProofPoint;
+}
+
+export { ProofPointIssueResult };
