+++ src/proofPointValidateResult.ts
@@ -1,0 +1,9 @@
+import { ProofPointStatus } from "./proofPointStatus";
+
+interface ProofPointValidateResult {
+  isValid: boolean;
+  statusCode: ProofPointStatus;
+  statusMessage: string | null;
+}
+
+export { ProofPointValidateResult };
