+++ test/contractProofPointRegistryStorage1.js
