+++ test/contractProofPointRegistry_v2.js
